+++ modules/graphics/builtin/include/graphics_object_deleteBuiltin.hpp
@@ -7,33 +7,14 @@
 // SPDX-License-Identifier: LGPL-3.0-or-later
 // LICENCE_BLOCK_END
 //=============================================================================
-#include "GOScalarDoubleProperty.hpp"
+#pragma once
 //=============================================================================
-namespace Nelson {
+#include "ArrayOf.hpp"
 //=============================================================================
-void
-GOScalarProperty::data(double x)
-{
-    at(0) = x;
-}
+namespace Nelson::GraphicsGateway {
 //=============================================================================
-double
-GOScalarProperty::data()
-{
-    return at(0);
-}
-//=============================================================================
-std::wstring
-GOScalarProperty::toWideString()
-{
-    double value = at(0);
-    std::wstring msg;
-    if (std::fabs((double)(int)(value)-value) < std::numeric_limits<double>::epsilon()) {
-        msg = std::to_wstring((int)value);
-    } else {
-        msg = std::to_wstring(value);
-    }
-    return msg;
-}
+ArrayOfVector
+graphics_object_deleteBuiltin(int nLhs, const ArrayOfVector& argIn);
 //=============================================================================
 }
+//=============================================================================